Add initial ontcluster library
Mostly written by Codex using the following prompt, but then refined and edited significantly to remove unnecessary complexity and fix some bugs.

Make a new python library using BioPython called OntCluster suitable for use in my jupyter notebooks or on the command line. It should use BioPython to take a reference sequence, a FASTQ file and a noise threshold parameter and output a set of FASTQ files containing exactly all the input reads but clustered into distinct sequences. This is for sequencing PCR amplicons of mixed samples. In order to make this work well you will need to build it up from simpler components, each of which will have small unit tests demonstrating it's correctness. Prefer clear and simple code, being careful to avoid unnecessary complexity or extra lines of code. Performance is also important and the reads may be too big to store all in memory at once, so the FASTQ input and output files should be processed in a streaming fashion. Put all tests in a new 'tests' subdirectory using filenames that start with the name of the file being tested.

First write a MakePileup function which takes a reference sequence and iterator of SeqRecord objects ("reads") and generates a Pileup data structure which consists of the reference sequence and an array of variations indexed by reference position. The variations includes a count of A, C, T, G and N, as well as "DEL" for deletion and Insertion data structures (eg. Insertion('CAG')). MakePileup works by taking each record in the input FASTQ file and using PairwiseAligner to align it to the reference sequence, then walking the alignment to update the Pileup. Reads which don't align to the reference sequence at all should just be counted as unaligned and otherwise ignored. Write some small unit tests for MakePileup - testing basic cases on small example sequences including SNPs, deletions and a few different insertions.

Once you are confident that MakePileup works correctly, write a function called DenoisePileup which takes one Pileup and a noise threshold parameter and generates a new Pileup containing only the variations with frequency above the noise threshold. For example if the threshold is 0.05 and at one position there are 90 'A's, 7 'AG' insertions, one 'AGC' insertion and 2 'G's, the result should contain only the 90 'A's and 7 'AG' insertions. Write some small unit tests for this which demonstrates its correctness.

Now write a function called ReadPileupIntersect which given a reference Pileup and read SeqRecord returns a new Seq which contains only the variations from read which are present in the consensus Pileup. It does this by doing a PairwiseAlignment of the read to the Pileup's reference. If the alignment does not cover the reference sequence, then return failure. Otherwise walk the alignment and build up a Seq from the Pileup's reference and possible variations. Add helper functions on the Pileup class as needed to make this clear and straightforward. Write several unit tests for different configuations, making the unit tests as clear as possible.

Now put this all together by processing the input FASTQ files in three passes (using three separate helper functions). First send all reads to MakePileup, sending the result to DenoisePileup and allowing the original Pileup to be garbage collected. Output how many reads were processed and how many ignored as unaligned, as well as how many variations exist in the denoised pileup. Then loop over the input reads again selecting only the reads which are at least as long as the reference sequence. For each of these, use ReadPileupIntersect to build up a set of 'cluster' Seq objects - a dictionary of sequences to numbers. The first Seq created becomes cluster 1. Then for each Seq generated, check whether it exists exactly in the set of clusters, and if so just skip it. Otherwise create a new cluster entry with the next index. Output the total number of clusters. Now create FASTQ files for each cluster - matching the input FASTQ base filename with "-C1.fastq" appended for cluster 1 etc. Then for the third pass over all the input reads, do a score-only PairwiseAlignment to each cluster. The read is asigned to whichever cluster gets the highest score. Append the read to appropriate FASTQ file. Output the number of reads written to each output file. Write one integration test for this whole setup using some simple generated input.